Revert "floats for icc check"
This reverts commit f9c9b12756acb4a644785147360b92fbf06fe3e3.

+++ stencil.c
@@ -6,9 +6,9 @@
 // Define output file name
 #define OUTPUT_FILE "stencil.pgm"
 
-void stencil(const int nx, const int ny, float *  image, float *  tmp_image);
-void init_image(const int nx, const int ny, float *  image, float *  tmp_image);
-void output_image(const char * file_name, const int nx, const int ny, float *image);
+void stencil(const int nx, const int ny, double *  image, double *  tmp_image);
+void init_image(const int nx, const int ny, double *  image, double *  tmp_image);
+void output_image(const char * file_name, const int nx, const int ny, double *image);
 double wtime(void);
 
 int main(int argc, char *argv[]) {
@@ -25,8 +25,8 @@ int main(int argc, char *argv[]) {
   int niters = atoi(argv[3]);
 
   // Allocate the image
-  float *image = malloc(sizeof(float)*nx*ny);
-  float *tmp_image = malloc(sizeof(float)*nx*ny);
+  double *image = malloc(sizeof(double)*nx*ny);
+  double *tmp_image = malloc(sizeof(double)*nx*ny);
 
   // Set the input image
   init_image(nx, ny, image, tmp_image);
@@ -50,7 +50,7 @@ int main(int argc, char *argv[]) {
 }
 
 //called 200 times
-void stencil(const int nx, const int ny, float * restrict image, float * restrict tmp_image) {
+void stencil(const int nx, const int ny, double * restrict image, double * restrict tmp_image) {
   for (int i = 0; i < nx; ++i) {
     for (int j = 0; j < ny; ++j) {
       tmp_image[j+i*nx] = image[j+i*nx] * 0.6;
@@ -62,7 +62,7 @@ void stencil(const int nx, const int ny, float * restrict image, float * restric
   }
 }
 /*
-void stencil(const int nx, const int ny, float * image, float * tmp_image) {
+void stencil(const int nx, const int ny, double * image, double * tmp_image) {
   const int max = nx*ny;
   const int modnix = (max%nx)-1;
   for(int z = 0; z <= max; z++){
@@ -75,7 +75,7 @@ void stencil(const int nx, const int ny, float * image, float * tmp_image) {
 }
 */
 // Create the input image
-void init_image(const int nx, const int ny, float *  image, float *  tmp_image) {
+void init_image(const int nx, const int ny, double *  image, double *  tmp_image) {
   // Zero everything
   const int max = nx*ny;
   for (int z = 0; z < max; z++){
@@ -97,7 +97,7 @@ void init_image(const int nx, const int ny, float *  image, float *  tmp_image) 
 }
 
 // Routine to output the image in Netpbm grayscale binary image format
-void output_image(const char * file_name, const int nx, const int ny, float *image) {
+void output_image(const char * file_name, const int nx, const int ny, double *image) {
 
   // Open output file
   FILE *fp = fopen(file_name, "w");
@@ -112,7 +112,7 @@ void output_image(const char * file_name, const int nx, const int ny, float *ima
   // Calculate maximum value of image
   // This is used to rescale the values
   // to a range of 0-255 for output
-  float maximum = 0.0;
+  double maximum = 0.0;
   for (int j = 0; j < ny; ++j) {
     for (int i = 0; i < nx; ++i) {
       if (image[j+i*ny] > maximum)
